requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
websockets>=12.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.0

# LLM / agents
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langgraph>=0.1.0

# Vector store
weaviate-client>=4.6.0

# Cache / memory
redis[hiredis]>=5.0.0

# RAG / document processing
pypdf>=4.2.0
sentence-transformers>=3.0.0
tiktoken>=0.7.0

# Utilities
rich>=13.7.0
typer>=0.12.0
schedule>=1.2.0
numpy>=1.26.0
pandas>=2.2.0
python-multipart>=0.0.9
jinja2>=3.1.4
